+++ docs/changelog.rst
@@ -7,6 +7,7 @@ Unreleased
 * Default HOST to localhost for postgres databases. https://github.com/jazzband/django-dbbackup/issues/520
 * Add PostgreSQL Schema support by @angryfoxx in https://github.com/jazzband/django-dbbackup/pull/507
 * Fix restore of database from S3 storage by reintroducing inputfile.seek(0) to utils.uncompress_file
+* Add warning for filenames with slashes in them
 * Fix bug where dbbackup management command would not respect settings.py:DBBACKUP_DATABASES
 * Remove usage of deprecated 'get_storage_class' function in newer Django versions
 * Add support for new STORAGES (Django 4.2+) setting under the 'dbbackup' alias